Add a governance policy
As the eRegs community grows, we have need for more structure. With this
change, I'm proposing an initial governance policy around how changes are
made. It more or less codifies what we've done thus far. Let's follow its
suggestion: we'll leave this PR open for two weeks for discussion. If all is
well, we'll merge at the end.